changing 'PlayerChoices' to 'DetermineWinner' since that's its only function now
